Update protocol.py
To avoid: DeprecationWarning: The default dtype for empty Series will be 'object' instead of 'float64' in a future version. Specify a dtype explicitly to silence this warning.
  for asset, position in self.positions.items()

+++ catalyst/protocol.py
@@ -190,7 +190,7 @@ class Portfolio(object):
                 asset_multiplier(asset)
             )
             for asset, position in self.positions.items()
-        })
+        }, dtype='float64')
         return position_values / self.portfolio_value
